+++ src/inmanta/config.py
@@ -215,18 +215,26 @@ class Option(object):
     :param documentation: the documentation for this option
     :param validator: a function responsible for turning the string representation of the option into the correct type.
         Its docstring is used as representation for the type of the option.
+    :param predecessor_option: The Option that was depricated in favour of this option.
     """
 
-    def __init__(self, section, name, default, documentation, validator=is_str):
+    def __init__(self, section, name, default, documentation, validator=is_str, predecessor_option=None):
         self.section = section
         self.name = _normalize_name(name)
         self.validator = validator
         self.documentation = documentation
         self.default = default
+        self.predecessor_option = predecessor_option
         Config.register_option(self)
 
     def get(self):
         cfg = Config._get_instance()
+        if self.predecessor_option:
+            has_depricated_option = cfg.has_option(self.predecessor_option.section, self.predecessor_option.name)
+            has_new_option = cfg.has_option(self.section, self.name)
+            if has_depricated_option and not has_new_option:
+                LOGGER.warning("Config option %s is depricated. use %s instead." % (self.predecessor_option.name, self.name))
+                return self.predecessor_option.get()
         out = cfg.get(self.section, self.name, fallback=self.get_default_value())
         return self.validate(out)